Extract shard I/O helpers into shardIo module
Move the seven sharded-data file helpers and their three constants
out of `utils.py` into a new `shardIo` module:

- ALLOWED_DATASET_TYPES, SEQNUM_PADDING, SHARDED_DATA_TEMPLATE
- writeMetadataShard / writeExpRecordMetadataShard — write per-image
  metadata shards into the shared directory the
  TimedMetadataServer reads
- writeDataShard / createFilenameForDataShard — write per-image data
  shards (binnedImage, rawNoises) into the same directory
- getShardedData — read back the matching shards for a given
  dayObs/seqNum/dataset, with timeout-and-retry semantics
- getGlobPatternForShardedData — the wildcard pattern the reader
  uses to find shards on disk
- getShardPath — instrument-specific routing into the right
  per-camera shard directory on the LocationConfig

These all together form the read/write side of the worker → server
shard hand-off, so they belong in one module.

The new module imports `NumpyEncoder` from `parsers` (for the
JSON-encoder argument used when writing) and `isFileWorldWritable`
from `predicates` (for the post-write permission fixup), and uses
the FILTER_DELIMITER constant from `obs_lsst` for the LATISS filter
splitting in `writeExpRecordMetadataShard`. The TYPE_CHECKING import
of `LocationConfig` still points at `.utils` for now and will be
re-pointed once `LocationConfig` itself moves.

All callers across the package and the test suite have been updated
to import from `shardIo` instead of `utils`. The previously-unused
imports of `glob`, `time`, `uuid`, `FILTER_DELIMITER`, `NumpyEncoder`
and `isFileWorldWritable` have been removed from `utils.py`.

Verified:
- mypy and flake8 clean.
- The full unit test suite (185 tests) passes.
- Every dependent in-package module imports cleanly.

Written by Claude Code 4.6 Opus (high-effort).
