+++ python/lsst/rubintv/production/aos.py
@@ -73,8 +73,9 @@ from .aosUtils import (
 )
 from .formatters import getRubinTvInstrumentName, makePlotFile
 from .redisUtils import RedisHelper, _extractExposureIds
+from .shardIo import writeExpRecordMetadataShard, writeMetadataShard
 from .uploaders import MultiUploader
-from .utils import logDuration, writeExpRecordMetadataShard, writeMetadataShard
+from .utils import logDuration
 
 if TYPE_CHECKING:
     from lsst.daf.butler import Butler, DimensionRecord
