+++ python/lsst/rubintv/production/processingControl.py
@@ -55,14 +55,9 @@ from .payloads import Payload, pipelineGraphToBytes
 from .podDefinition import PodDetails, PodFlavor
 from .predicates import isCalibration, isWepImage, raiseIf, runningCI
 from .redisUtils import ExposureProcessingInfo, RedisHelper
+from .shardIo import getShardPath, writeExpRecordMetadataShard, writeMetadataShard
 from .timing import BoxCarTimer
-from .utils import (
-    LocationConfig,
-    getExpIdOrVisitId,
-    getShardPath,
-    writeExpRecordMetadataShard,
-    writeMetadataShard,
-)
+from .utils import LocationConfig, getExpIdOrVisitId
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
