+++ python/lsst/rubintv/production/timedServices.py
@@ -38,8 +38,9 @@ import matplotlib.pyplot as plt
 
 from .parsers import sanitizeNans
 from .predicates import hasDayRolledOver, isFileWorldWritable, raiseIf
+from .shardIo import writeMetadataShard
 from .uploaders import MultiUploader
-from .utils import logDuration, writeMetadataShard
+from .utils import logDuration
 
 try:
     from lsst_efd_client import EfdClient  # noqa: F401 just check we have it, but don't use it
